Messages rectored to a component
